requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
pydantic==2.5.0
uvicorn==0.24.0
streamlit==1.28.1
requests==2.31.0
PyJWT==2.8.0
